rename to pitwall slice
