+++ src/lib/redux/slices/index.ts
@@ -1,6 +1,6 @@
 export * from "./authSlice";
 export * from "./preferencesSlice";
 export * from "./sessionSlice";
-export * from "./pitwallSessionSlice";
+export * from "./pitwallSlice";
 export * from "./standingsSlice";
 export * from "./telemetrySlice";
